+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 ## Features
 
+* Extract default migration filenames + add getter
 * Add use statement for EPB annotations and switch annotation from @Migrate to @EPB\Migrate in migrations template
 
 # Version 7.0.1